キー、マウス入力の追加

+++ Kawabata_ShootingGame/Kawabata_ShootingGame/StraightBullets.cpp
@@ -19,5 +19,5 @@ void StraightBullets::UpDate()
 //•`‰æ
 void StraightBullets::Draw()
 {
-	DrawCircle(GetLocation().x, GetLocation().y, GetRadius(), GetColor(255, 0, 0));
+	DrawCircle(static_cast<int>(GetLocation().x), static_cast<int>(GetLocation().y), static_cast<int>(GetRadius()), GetColor(255, 0, 0));
 }
